requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
comtypes==1.4.3
cryptography==40.0.2
dataclasses==0.8
future==1.0.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
idna==3.10
importlib-metadata==4.8.3
iso8601==1.1.0
jwt==1.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
numpy==1.19.5
opencv-python==4.5.5.64
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pycparser==2.21
PyJWT==1.7.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyserial==3.5
PySide2==5.15.2.1
pywifi==1.1.12
PyYAML==6.0.1
requests==2.27.1
scipy==1.5.4
serial==0.0.97
shiboken2==5.15.2.1
six==1.17.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing_extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.6.0